refactor(wayland): move different indevs to their own file
