+++ src/drivers/wayland/lv_wl_keyboard.h
@@ -1,0 +1,46 @@
+/**
+ * @file lv_wayland_keyboard.h
+ *
+ */
+
+
+#ifndef LV_WL_KEYBOARD_H
+#define LV_WL_KEYBOARD_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*********************
+ *      INCLUDES
+ *********************/
+
+#include "../../indev/lv_indev.h"
+#include "../../indev/lv_indev_gesture.h"
+#if LV_USE_WAYLAND
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ * GLOBAL PROTOTYPES
+ **********************/
+
+lv_indev_t * lv_wayland_keyboard_create(void);
+
+/**********************
+ *      MACROS
+ **********************/
+
+#endif /* LV_USE_WAYLAND */
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*LV_WL_KEYBOARD_H*/
