+++ src/drivers/wayland/lv_wl_touch.h
@@ -1,0 +1,48 @@
+
+
+/**
+ * @file lv_wayland_touch.h
+ *
+ */
+
+
+#ifndef LV_WL_TOUCH_H
+#define LV_WL_TOUCH_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*********************
+ *      INCLUDES
+ *********************/
+
+#include "../../indev/lv_indev.h"
+#include "../../indev/lv_indev_gesture.h"
+#if LV_USE_WAYLAND && LV_USE_GESTURE_RECOGNITION
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ * GLOBAL PROTOTYPES
+ **********************/
+
+lv_indev_t * lv_wayland_touch_create(void);
+
+/**********************
+ *      MACROS
+ **********************/
+
+#endif /* LV_USE_WAYLAND */
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*LV_WL_TOUCH_H*/
